Prepare API for MQTT
